Extracted the Component into a Razor Lib
