+++ YelpRandomFood/Pages/Index.razor
@@ -1,4 +1,4 @@
-﻿@using YelpRandomRestaurantFinder.Components
+﻿@using YelpRestaurantFinderComponent.Components
 @page "/"
 
 <PageTitle>Index</PageTitle>
