+++ YelpRandomFood/YelpRandomRestaurantFinder.csproj
@@ -12,6 +12,11 @@
     <PackageReference Include="GraphQL.Client" Version="4.0.1" />
     <PackageReference Include="GraphQL.Client.Serializer.SystemTextJson" Version="4.0.1" />
     <PackageReference Include="IP2Location.IPGeolocation" Version="8.7.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.ProtectedBrowserStorage" Version="5.0.0-rc.1.20451.17" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\YelpRestaurantFinderComponent\YelpRestaurantFinderComponent.csproj" />
   </ItemGroup>
 
 </Project>
