+++ YelpRandomRestaurantFinder.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YelpRandomRestaurantFinder", "YelpRandomFood\YelpRandomRestaurantFinder.csproj", "{95EB29ED-2587-4F3E-98F4-D700044C40EB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YelpRestaurantFinderComponent", "YelpRestaurantFinderComponent\YelpRestaurantFinderComponent.csproj", "{C59DB7AF-BA70-4EF3-B58B-6D40F4E81BDA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{95EB29ED-2587-4F3E-98F4-D700044C40EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{95EB29ED-2587-4F3E-98F4-D700044C40EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{95EB29ED-2587-4F3E-98F4-D700044C40EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C59DB7AF-BA70-4EF3-B58B-6D40F4E81BDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C59DB7AF-BA70-4EF3-B58B-6D40F4E81BDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C59DB7AF-BA70-4EF3-B58B-6D40F4E81BDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C59DB7AF-BA70-4EF3-B58B-6D40F4E81BDA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
